- feat: refactor: migrar projeto para Clean Architecture e desacoplar camada Web
- Implementada separação de camadas em Core (Domain/Ports), Application (Use Cases) e Infrastructure.
- Criados DTOs de Request e Response para as entidades Cliente e Produto, garantindo o isolamento do modelo de domínio na API.
- Migrada a lógica de negócio dos Controllers para Use Cases dedicados.
- Removidos arquivos redundantes e códigos não utilizados (limpeza de pacotes).
- Inversão de dependência aplicada via interfaces (Ports) para persistência.
- Atualizada a documentação técnica (DOCUMENTACAO.md) e o checklist de progresso (STEPBYSTEP.md).

+++ src/main/java/galleriabank/compras/infrastructure/web/dtos/response/PedidoResponseDTO.java
@@ -1,6 +1,7 @@
 package galleriabank.compras.infrastructure.web.dtos.response;
 
 import galleriabank.compras.core.domain.Pedido;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -12,7 +13,7 @@ public record PedidoResponseDTO(
         Long clienteId,
         String clienteNome,
         List<ProdutoResponseDTO> produtos,
-        double total
+        BigDecimal total
 ) {
     public static PedidoResponseDTO fromEntity(Pedido pedido) {
         var listaProdutos = pedido.getProdutos().stream()
